schedulerService/requirements.txt
python-dotenv 
requests
schedule